chore(ruff): adopt high-signal rule families + fix findings
Expand ruff select beyond the bug-catching core (E4/E7/E9,F,I,B,UP) with
low-false-positive families that fit this codebase's clean / no-dead-code
values, measured against the actual tree before adopting:

  +C4 +PIE +ISC +RSE +FLY  near-zero-hit regression ratchets
  +RUF                     mutable-class-default, unused-unpacked-var, unsorted __all__
  +PERF                    manual loops that should be comprehensions/extend
  +SIM                     collapsible-if, ternary, suppressible-exception
  +PTH                     pathlib over os.*

ignore = RUF001-003 (false positives on intentional Cyrillic), ISC001
(conflicts with the formatter). per-file-ignore PERF401 for tests/ (perf
lint is irrelevant to fixtures). Deliberately NOT adopted: D/ANN (basedpyright
owns typing; docstrings not mandated), S101/PLR/TRY003 (noise), DTZ (fights
the intentional naive-local calendar values — left as its own decision).

Findings fixed (not blanket-suppressed):
- PERF401: 2 SQL builders -> [seed, *(gen), tail]; 2 filter-loops -> .extend;
  shoes parse -> comprehension
- RUF012: Config.Zones.COEFF, Config.Drift.THRESHOLD_BY_SPORT, a test fake -> ClassVar
- SIM105: try/except/pass -> contextlib.suppress(Exception)
- SIM108/102: ternary / collapsed nested-if
- PTH: os.replace/os.chmod/open() -> Path.replace/.chmod/.open (drop unused import os)

Two deliberate suppressions with rationale:
- metric_services._row_get  # noqa: SIM118 (polymorphic Row|dict; .keys() is the safe contract)
- test_settings  # noqa: FLY002 (multi-line env list reads better than one literal)

just check (ruff + basedpyright reportAny=0 + vulture) green; 389 passed; warm MCP smoke ok.

+++ src/mcp_strava/application/metric_services.py
@@ -239,7 +239,9 @@ def _fact_status(row: ActivityMetricFactRow) -> dict[str, Any]:
 def _row_get(row: object, key: str, default: object = None) -> object:
     if row is None:
         return default
-    if hasattr(row, "keys") and key in row.keys():  # type: ignore[union-attr]
+    # SIM118 suppressed: row is a polymorphic DuckDB Row | dict; explicit .keys()-membership
+    # is the safe contract — `key in row` is not guaranteed key-membership on Row objects.
+    if hasattr(row, "keys") and key in row.keys():  # type: ignore[union-attr]  # noqa: SIM118
         return row[key]  # type: ignore[index]
     if isinstance(row, dict):
         return row.get(key, default)
@@ -367,7 +369,7 @@ def _activity_payload(
         "elevation_m": _activity_value(row, "elevation_m"),
         "trimp": _activity_value(row, "trimp"),
         "avg_hr": summary.get("average_heartrate"),
-        "max_hr": int(round(summary["max_heartrate"])) if summary.get("max_heartrate") else None,  # type: ignore[arg-type]
+        "max_hr": round(summary["max_heartrate"]) if summary.get("max_heartrate") else None,  # type: ignore[arg-type]
         "time_in_hr_zones_min": _zone_minutes(row),
         "hr_recovery_pauses": int(row["hr_recovery_pause_count"] or 0),  # type: ignore[arg-type]
         "hr_recovery_total_rest_sec": int(row["hr_recovery_total_rest_sec"] or 0),  # type: ignore[arg-type]
